added a mongodb and created user model
