+++ README.md
@@ -1,0 +1,60 @@
+<!-- PROJECT LOGO -->
+<br />
+<p align="center">
+  <!-- <a href="https://github.com/garzeah/restaurant-tinder">
+    <img src="src/assets/images/desktopLogo.png" alt="Decipher Logo">
+  </a> -->
+
+  <h3 align="center">Decipher</h3>
+
+  <p align="center">
+    Helps people overcome language barriers.
+  </p>
+</p>
+
+<!-- ABOUT THE PROJECT -->
+
+## About The Project
+
+Decipher allows users to chat in different languages and overcome language barriers!
+
+### Built With
+
+- React
+- Node
+- Express
+- Mongo
+
+### Installation
+
+<!-- 1. Clone the repo
+   ```sh
+   git clone https://github.com/github_username/repo_name.git
+   ```
+2. Install NPM packages
+   ```sh
+   npm install
+   ```
+3. Start the server
+   ```sh
+   npm start
+   ``` -->
+
+1. Create a keys.js file in the config folder
+2. Paste the following template
+   ```sh
+   module.exports = {
+   mongoURI: "",
+   };
+   ```
+3. Enter your MongoURI (can get one from MongoDB Atlas)
+
+<!-- CONTACT -->
+
+## Contact
+
+Andy Garcia: acgarzeah@gmail.com
+
+Project Link: https://github.com/garzeah/decipher
+
+Website: N/A
