--- index.js
@@ -1,11 +1,0 @@
-const express = require("express");
-const app = express();
-
-app.get("/", (req, res) => {
-  res.send({ hi: "there" });
-});
-
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => {
-  console.log(`Server has started on port ${PORT}`);
-});
