+++ models/User.js
@@ -1,0 +1,10 @@
+const mongoose = require("mongoose");
+const { Schema } = mongoose;
+
+const userSchema = new Schema({
+  name: String,
+  email: String,
+  password: String,
+});
+
+mongoose.model("users", userSchema);
